Added husky to support linting as posts are committed
