+++ package.json
@@ -14,6 +14,7 @@
     "glob-promise": "^4.2.2",
     "globby": "^7.1.1",
     "gray-matter": "^3.1.1",
+    "husky": "^9.1.7",
     "js-yaml": "^3.10.0",
     "markdown-spellcheck": "^1.3.1",
     "markdown-to-txt": "^2.0.0",
@@ -32,12 +33,14 @@
   },
   "scripts": {
     "lint": "node lintPosts.js",
+    "lint-commit": "node lintCommit.js",
     "compute-embeddings": "node scripts/generate-related/compute-embeddings.js",
     "generate-related": "node scripts/generate-related/blog-metadata.js",
     "remove-unused-images": "node scripts/images/remove-images.js",
     "spellcheck": "mdspell \"**/ceberhardt/_posts/*.md\" --en-gb -a -n -x -t",
     "style": "sass --no-source-map --style=compressed scss/style.scss style.css",
-    "scripts": "uglifyjs scripts/initialise-menu.js scripts/jquery-1.9.1.js scripts/jquery.jscroll-2.2.4.js scripts/load-clap-count.js scripts/elapsed.js scripts/graft-studio/header-scroll.js scripts/graft-studio/jquery.mmenu.all.js scripts/graft-studio/jquery.matchHeight.js node_modules/applause-button/dist/applause-button.js node_modules/cookieconsent/build/cookieconsent.min.js -o script.js"
+    "scripts": "uglifyjs scripts/initialise-menu.js scripts/jquery-1.9.1.js scripts/jquery.jscroll-2.2.4.js scripts/load-clap-count.js scripts/elapsed.js scripts/graft-studio/header-scroll.js scripts/graft-studio/jquery.mmenu.all.js scripts/graft-studio/jquery.matchHeight.js node_modules/applause-button/dist/applause-button.js node_modules/cookieconsent/build/cookieconsent.min.js -o script.js",
+    "prepare": "husky"
   },
   "homepage": "http://blog.scottlogic.com",
   "private": true
